backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.16.4
pydantic[email]==2.11.7
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
httpx==0.28.1
redis[hiredis]==5.2.1
celery[redis]==5.5.3
stripe==12.4.0
sendgrid==6.12.4
boto3==1.39.14
python-multipart==0.0.20
pillow==11.3.0
pytest==8.4.1
pytest-asyncio==1.1.0
factory-boy==3.3.3
ruff==0.12.7
mypy==1.17.0
black==25.1.0
prometheus-fastapi-instrumentator==7.1.0
